+++ src/3_branch.js
@@ -1065,6 +1065,7 @@ if (WEB_BUILD) { // jshint undef:false
  *     description: 'The Branch demo app!',
  *     openAppButtonText: 'Open',         // Text to show on button if the user has the app installed
  *     downloadAppButtonText: 'Download', // Text to show on button if the user does not have the app installed
+ *     sendLinkText: 'Send Link',         // Text to show on desktop button to allow users to text themselves the app
  *     phonePreviewText: '+44 9999-9999', // The default phone placeholder is a US format number, localize the placeholder number with a custom placeholder with this option
  *     iframe: true,                      // Show banner in an iframe, recomended to isolate Branch banner CSS
  *     showiOS: true,                     // Should the banner be shown on iOS devices?
@@ -1115,6 +1116,7 @@ if (WEB_BUILD) { // jshint undef:false
 			description: options['description'] || '',
 			openAppButtonText: options['openAppButtonText'] || 'View in app',
 			downloadAppButtonText: options['downloadAppButtonText'] || 'Download App',
+			sendLinkText: options['sendLinkText'] || 'Send Link',
 			phonePreviewText: options['phonePreviewText'] || '(999) 999-9999',
 			iframe: typeof options['iframe'] == 'undefined' ? true : options['iframe'],
 			showiOS: typeof options['showiOS'] == 'undefined' ? true : options['showiOS'],