+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to the Branch Web SDK will be documented here.
 The Branch Web SDK adheres to [Semantic Versioning](http://semver.org/).
 
+## [VERSION] - unreleased
+
 ## [1.4.1] - 2015-04-29
 
 ## [1.4.0] - 2015-04-25