+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to the Branch Web SDK will be documented here.
 The Branch Web SDK adheres to [Semantic Versioning](http://semver.org/).
 
 ## [VERSION] - unreleased
+### New Functionality
+- Adds data_parsed to branch.init() and referring_data_parsed to branch.setIdentity()
 
 ## [1.4.2] - 2015-05-06
 - Fixes bug in link creation without data