+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The Branch Web SDK adheres to [Semantic Versioning](http://semver.org/).
 ## [VERSION] - unreleased
 ### New Functionality
 - Adds smart banner position option, with possible values: 'top', or 'bottom'
+- Adds data_parsed to branch.init() and referring_data_parsed to branch.setIdentity()
 
 ## [1.4.2] - 2015-05-06
 - Fixes bug in link creation without data